pyproject.toml
[tool.poetry]
name = "owyn-voice-api"
version = "0.1.0"
description = ""
authors = ["Owyn Richen <owynrichen@gmail.com>"]
readme = "README.md"
package-mode = false

[tool.poetry.dependencies]
python = "^3.11"
fastapi = {extras = ["standard"], version = "^0.115.6"}
suno-bark = {git = "https://github.com/suno-ai/bark.git"}
melotts = { git = "https://github.com/owynrichen/MeloTTS", branch = "bugfix/update_gruut_2.4" }
openvoice = { git = "https://github.com/owynrichen/OpenVoice", branch = "feature/add_pyproject" }
transformers = "^4.47.1"
scipy = "^1.15.0"
unidic = "^1.1.0"
unidic-lite = "^1.0.8"

# https://github.com/lucaspar/poetry-torch
torch = {version = "^2.5.1", extras = ["cuda"]}
torchaudio = {version = "^2.5.1", extras = ["cuda"]}
xformers = {version = "^0.0.29.post1", extras = ["cuda"]}
cloudflare = "^4.0.0"
elevenlabs = "^1.50.4"

[tool.poetry.group.cuda]
    optional = true

    # https://github.com/pytorch/pytorch/blob/main/RELEASE.md#release-compatibility-matrix
    [tool.poetry.group.cuda.dependencies]
        torch       = { version = "^2.5.1", source = "pytorch-cuda", markers = "extra=='cuda' and extra!='cpu'" }
        torchvision = { version = "^0.20", source = "pytorch-cuda", markers = "extra=='cuda' and extra!='cpu'" }
        torchaudio  = { version = "^2.5.1", source = "pytorch-cuda", markers = "extra=='cuda' and extra!='cpu'" }
        xformers    = { version = "^0.0.29", source = "pytorch-cuda", markers = "extra=='cuda' and extra!='cpu'" }

[tool.poetry.extras]
    cpu  = ["torch", "torchvision", "torchaudio"]
    cuda = ["torch", "torchvision", "torchaudio"]

[[tool.poetry.source]]
    name     = "pytorch-cuda"
    priority = "explicit"
    url      = "https://download.pytorch.org/whl/cu124"

[[tool.poetry.source]]
    name     = "pytorch-cpu"
    priority = "explicit"
    url      = "https://download.pytorch.org/whl/cpu"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
